+++ src/models/topic.ts
@@ -1,13 +1,10 @@
 import { Data, type Require as Enforced } from "dataclass";
-import * as fs from "fs";
+
 import * as vscode from "vscode";
 import { KafkaTopicOperation } from "../authz/types";
 import { ConnectionType } from "../clients/sidecar";
 import { IconNames } from "../constants";
 import { CustomMarkdownString } from "./main";
-import { getTopicViewProvider } from "../viewProviders/topics";
-import { getSidecar } from "../sidecar";
-import { registerCommandWithLogging } from "../commands";
 import { ConnectionId, IResourceBase, isCCloud } from "./resource";
 
 /** Main class representing Kafka topic */
@@ -135,51 +132,4 @@ function createKafkaTopicTooltip(
   }
 
   return tooltip;
-}
-async function produceMessageFromFile() {
-  const options: vscode.OpenDialogOptions = {
-    canSelectMany: false,
-    openLabel: "Select JSON file",
-    filters: {
-      "JSON files": ["json"],
-    },
-  };
-
-  const fileUri = await vscode.window.showOpenDialog(options);
-  if (fileUri && fileUri[0]) {
-    const filePath = fileUri[0].fsPath;
-    const message = JSON.parse(fs.readFileSync(filePath, "utf-8"));
-    const topic = getTopicViewProvider().selectedTopic;
-
-    if (!topic) {
-      vscode.window.showErrorMessage("No topic selected.");
-      return;
-    }
-
-    const sidecar = await getSidecar();
-    const clusterId = topic.clusterId;
-    const connectionId = topic.connectionId;
-
-    const recordsApi = sidecar.getRecordsV3Api(clusterId, connectionId);
-
-    try {
-      await recordsApi.produceRecord({
-        topic_name: topic.name,
-        cluster_id: clusterId,
-        ProduceRequest: {
-          value: message as any,
-          //patching with any here, need to understand WHY this is necessary
-        },
-      });
-      vscode.window.showInformationMessage(`Message produced to topic ${topic.name}`);
-    } catch (error: any) {
-      vscode.window.showErrorMessage(`Failed to produce message: ${error.message}`);
-    }
-  }
-}
-
-export function registerKafkaClusterCommands(): vscode.Disposable[] {
-  return [
-    registerCommandWithLogging("confluent.organizations.produce.fromFile", produceMessageFromFile),
-  ];
 }
